refactor: move docs and resources objects to _global.py file
They required imports that were directly exposed to the user (e.g. import crappy.uint8)
